--- Chapter 10 Programs/Pg281-10-03-RemoveDash/Main.java
@@ -1,15 +1,0 @@
-// Code written as part of the Java-Methods-Program-AP-Comp-A-2021-2022 repository on GitHub.
-package com.company;
-
-public class Main {
-
-    public static void main(String[] args) {
-        String stringSecurityNum = "987-65-4321";
-        System.out.println(removeDashes(stringSecurityNum));
-    }
-
-    public static String removeDashes(String socialSecurityNum) {
-        return socialSecurityNum.replace("-", "");
-    }
-}
-//
