+++ src/main/java/chapter09/Chapter 9.md
@@ -1,0 +1,14 @@
+# AP Computer Science A
+## Chapter 9 - Implementing Classes and Using Objects
+
+| **Page** | **Problem** | **Package**               |
+|----------|-------------|---------------------------|
+| 241      |             | Snackbar                  |
+| 259      | 10a         | chapter09.rectangle.Rectangle                 |
+| 260      | 13          | TimeClass                 |
+| 261      | 14a<br>14b  | Coins<br>CoinsAndSnackbar |
+| 263      | 20          | GridLocation              |
+
+Problems taken from *Java Methods Object-Oriented Programming and Data Structures*.
+
+All code is written by Ethan Koh for the Java-Methods-Program-AP-Comp-A-2021-2022 repository on [GitHub](https://github.com/ArsiaC01/Java-Methods-Programs-AP-Comp-A-2021-2022).
