+++ src/main/java/chapter09/coinsandsnackbar/coin.gif
