+++ src/main/java/chapter10/Chapter 10.md
@@ -1,0 +1,13 @@
+# AP Computer Science A
+## Chapter 10 - Strings
+
+| **Page** | **Problem** | **Package**                |
+|----------|-------------|----------------------------|
+| 281      | 3           | RemoveDash                 |
+| 288      | 5a<br>5b    | CreditCardNumber           |
+| 289      | 14          | RemoveHtmlTag              |
+| 290      | 18a<br>18b  | Palindrome1<br>Palindrome2 |
+
+Problems taken from *Java Methods Object-Oriented Programming and Data Structures*.
+
+All code is written by Ethan Koh for the Java-Methods-Program-AP-Comp-A-2021-2022 repository on [GitHub](https://github.com/ArsiaC01/Java-Methods-Programs-AP-Comp-A-2021-2022).
